returned a possibility to have 64 bit hashes as this does not influence the performance.

+++ inc/G2DHashMapTrie.hpp
@@ -204,7 +204,7 @@ namespace uva {
                 inline void get_bucket_id(const T_M_Gram &gram, TShortId & bucket_idx) {
                     //Compute the hash value for the given M-gram, it must
                     //be the M-Gram id in the M-Gram data storage
-                    const uint32_t gram_hash = gram.hash();
+                    const uint64_t gram_hash = gram.hash();
                     LOG_DEBUG1 << "The " << gram.level << "-gram: " << tokensToString(gram)
                             << " hash is " << gram_hash << END_LOG;
